exchange fill_db scripts
