+++ .travis.yml
@@ -25,4 +25,4 @@ script:
   - flake8 ./ --max-line-length=130 --exclude=migrations --max-complexity 10
   - nohup python manage.py runserver &
   - sleep 10
-  - python manage.py local_fill_db
+  - python manage.py fill_db
